Restricted the pending deposit index to uint32 as the pending deposits container height is only 28

+++ contracts/contracts/strategies/NativeStaking/CompoundingValidatorManager.sol
@@ -983,7 +983,7 @@ abstract contract CompoundingValidatorManager is Governable, Pausable {
     struct PendingDepositProofs {
         bytes32 pendingDepositContainerRoot;
         bytes pendingDepositContainerProof;
-        uint40[] pendingDepositIndexes;
+        uint32[] pendingDepositIndexes;
         bytes[] pendingDepositProofs;
     }
